Fix: Refactoring model ExerciseContent
Put into model related "internal" methods and type
- ExerciseContentBlockSerializer.cs
- ExerciseContentBlock.cs
